requirements.txt
flask>=3.0
anthropic>=0.40
apscheduler>=3.10
python-dotenv>=1.0
pypdf>=4.0
pymupdf>=1.24
httpx>=0.27
psycopg2-binary>=2.9
